requirements.txt
Django>=5.1,<6.1
psycopg[binary]>=3.2
python-dotenv>=1.0
gunicorn>=22.0
